Reading of File complete

+++ BookDataCollector.hpp
@@ -13,12 +13,16 @@
 #include <vector>
 #include <map>
 #include <memory>
+#include <sstream>
+#include <fstream>
+#include <istream>
+#include <string>
 
 class BookDataCollector {
 private:
 public:
     BookDataCollector(std::vector<std::shared_ptr<Book>> &books, std::map<std::string, std::shared_ptr<Book>> &booksByTitle,
-                      std::map<uint64_t, std::shared_ptr<Book>> &booksByISBN);
+                      std::map<uint32_t, std::shared_ptr<Book>> &booksByISBN);
 };
 
 #endif /* BookDataCollector_hpp */
